+++ platform-emf/src/main/java/com/softicar/platform/emf/management/importing/EmfImportPopup.java
@@ -5,12 +5,12 @@ import com.softicar.platform.common.core.utils.CastUtils;
 import com.softicar.platform.dom.elements.popup.DomPopup;
 import com.softicar.platform.emf.EmfI18n;
 import com.softicar.platform.emf.management.EmfManagementMarker;
-import com.softicar.platform.emf.management.importing.analyze.EmfImportAnalyzeDiv;
 import com.softicar.platform.emf.management.importing.engine.EmfImportEngine;
 import com.softicar.platform.emf.management.importing.submit.EmfImportSubmitDiv;
 import com.softicar.platform.emf.management.importing.upload.EmfImportUploadDiv;
 import com.softicar.platform.emf.management.importing.upload.EmfImportViewUploadedDataDiv;
 import com.softicar.platform.emf.management.importing.variable.input.EmfImportVariablesInputDiv;
+import com.softicar.platform.emf.management.importing.variable.replace.EmfImportViewDataWithReplacementsDiv;
 import com.softicar.platform.emf.table.IEmfTable;
 import com.softicar.platform.emf.table.row.IEmfTableRow;
 
@@ -66,10 +66,10 @@ public class EmfImportPopup<R extends IEmfTableRow<R, P>, P, S> extends DomPopup
 		}
 	}
 
-	public void showAnalyzeDiv() {
+	public void showDataWithReplacementsDiv() {
 
 		removeChildren();
-		appendChild(new EmfImportAnalyzeDiv<>(this));
+		appendChild(new EmfImportViewDataWithReplacementsDiv<>(this));
 	}
 
 	public void showSubmitDiv() {
